updated code

+++ main.js
@@ -1,5 +1,5 @@
 const Discord = require('discord.js');
-const client = new Discord.Client();
+const client = new Discord.Client({ intents: ["GUILDS", "GUILD_MESSAGES"]});
 const prefix = 'f;';
 const deletedMessages = new Discord.Collection();
 const editedmessages = new Discord.Collection();
@@ -11,44 +11,50 @@ client.login(process.env.token1);
 // starting up tings
 client.once('ready', () => {
     console.log('real bot online now');
-    client.user.setActivity('called ur mum on big green pp');
-    client.user.setUsername(`[${prefix}] Tranquiity Gifter`);
-});
-
-
-// bans/kicks, memberjoins
-client.on('guildMemberAdd', user =>{
-    const channel = user.guild.channels.cache.find(channel => channel.name === "welcome");
-    if(!channel) return
-        channel.send(`Welcome to ${user.guild.name}, <@${user.id}>`);
-});
-client.on('guildMemberRemove', noobxd => {
-    const channel = noobxd.guild.channels.cache.find(channel => channel.name === "general");
-    channel.send(`xDDDD <@${noobxd.id}> got removed from this server lol mega pogchamp!!!`);
+    client.user.setStatus('dnd');
 });
 
 
 // commands
+// prefix help
+client.on('message', msg => {
+    if(msg.author.bot) return;
+    if(msg.content == prefix){
+        msg.reply('do the right command dog cunt ba\ni am ur father ogday');
+        msg.channel.send(`*${prefix}help*`)
+    }
+});
+// meme
+client.on('message', msg => {
+    if (msg.author.bot) return;
+    if (msg.content == `${prefix}meme`) {
+        const embed = new Discord.MessageEmbed()
+        .addFields({ name: ';)', value: '*you are the meme*', inline: true },)
+        .setThumbnail(msg.author.displayAvatarURL({dynamic:true}))
+        msg.channel.send(embed)
+    }
+});
 // avatar
 client.on('message', msg =>{
+    const kid = msg.mentions.members.first();
+    const args = msg.content.trim().split(/\s+/g);
+    const command1 = args.shift().toLowerCase();
     if (msg.author.bot) return;
-    if (msg.content === `${prefix}av`) {
-        if (msg.author.id === '420871370516201473') {
+    if (command1 === `${prefix}av`) {
+        if (!kid) {
             const embed = new Discord.MessageEmbed()
-                .setTitle('Looks Sexy 🥵 mm')
-                .setImage(msg.author.displayAvatarURL({dynamic:true}))
-                .setTimestamp()
+                .setTitle('this is you lmao')
+                .setImage(msg.author.displayAvatarURL({ size: 1024, dynamic: true }))
                 .setColor('#0CFF00');
             msg.channel.send(embed);
-            } else {
-                const embed = new Discord.MessageEmbed()
-                    .setTitle('Looks Sex- shit.')
-                    .setImage(msg.author.displayAvatarURL({dynamic:true}))
-                    .setTimestamp()
-                    .setColor('#0CFF00');
-                msg.channel.send(embed);
-            }
-    }
+        } else {
+            const embed = new Discord.MessageEmbed()
+                .setTitle('this is the person lmao')
+                .setImage(kid.user.displayAvatarURL({ size: 1024, dynamic: true }))
+                .setColor('#0CFF00');
+            msg.channel.send(embed);
+        }
+    }   
 });
 // snipe
 client.on('message', async message => {
@@ -65,12 +71,9 @@ client.on('message', async message => {
                 }, 2000);
             })
             const embed = new Discord.MessageEmbed()
-                .setAuthor(msg.author.tag, msg.author.avatarURL())
-                .setTimestamp()
-                .setDescription(msg.content);
-            message.channel.send(embed).catch(err => console.error(err)).then(E => {
+            message.channel.send().catch(err => console.error(err)).then(() => {
                 setTimeout(() => {
-                    E.delete()
+                   
                     message.delete().catch(E => console.error(E));
                 }, 10000);
             });
@@ -95,9 +98,9 @@ client.on('message', async message => {
                 .setAuthor(msg2.author.tag, msg2.author.avatarURL())
                 .setTimestamp()
                 .setDescription(msg2.content);
-            message.channel.send(embed2).catch(err => console.error(err)).then(E => {
+            message.channel.send(embed2).catch(err => console.error(err)).then(() => {
                 setTimeout(() => {
-                    E.delete()
+                   
                     message.delete().catch(E => console.error(E));
                 }, 10000);
             });
@@ -105,6 +108,7 @@ client.on('message', async message => {
 });
 // ping
 client.on('message', msg => {
+    if(msg.author.bot)return;
     if (msg.content === `${prefix}ping`) {
         msg.channel.send("pinging fag...").then(m => {
             setTimeout(() => {
@@ -118,7 +122,7 @@ client.on('message', msg => {
     if (msg.content === `${prefix}help`) {
         const embed = new Discord.MessageEmbed()
             .setTitle('this embed might help idk')
-            .setDescription(`Prefix[${prefix}] \n Cmds[av | snipe | editsnipe | ping | help | massping | membercount | mute | about]`)
+            .setDescription(`Prefix[${prefix}] \n Cmds[kick | id | av | snipe | editsnipe | ping | help | massping | membercount | mute | about]`)
             .setFooter('Made By Boomers Highness#1230')
             .setColor('#0CFF00')
             .setTimestamp();
@@ -128,17 +132,19 @@ client.on('message', msg => {
 // massping
 client.on('message', msg => {
     const kid = msg.mentions.members.first();
-
     if (msg.author.bot) return;
-    if (msg.content.startsWith(`${prefix}massping`) || msg.content.startsWith(`${prefix}mp`)) {
-        if (!kid) return msg.reply('mention someone dummmmie');
+    if (msg.content.startsWith(`${prefix}massping`) || msg.content.startsWith(`${prefix}mp`)) {   
+        if (!kid) return msg.reply('mention someone dummmmie').then(d =>{ 
+            setTimeout(() => {
+                d.delete();
+                msg.delete().catch(err => console.error(err)); 
+            }, 3000); 
+        });
         if (kid.id === '420871370516201473') {
             msg.channel.send(`no noob uno reverse ${msg.author}${msg.author}${msg.author}${msg.author}${msg.author}${msg.author}${msg.author}${msg.author}${msg.author}${msg.author}${msg.author}${msg.author}${msg.author}${msg.author}${msg.author}${msg.author}${msg.author}${msg.author}${msg.author}${msg.author}${msg.author}${msg.author}${msg.author}${msg.author}${msg.author}`).then((m) => {
-                setTimeout(() => {
-                    m.delete();
-                    msg.delete();
-                }, 5000)
-            })
+                m.delete();
+                msg.delete().catch(err => console.error(err));
+            });
         } else {
             msg.delete().catch(F => console.error(F));
             msg.channel.send(`ha someone evil did this too you sorry | ${kid} ${kid} ${kid} ${kid} ${kid} ${kid} ${kid} ${kid} ${kid} ${kid} ${kid} ${kid} ${kid} ${kid} ${kid} ${kid} ${kid} ${kid} ${kid} ${kid} ${kid} ${kid} ${kid} ${kid} ${kid} ${kid} ${kid} ${kid} ${kid} ${kid} ${kid} ${kid} ${kid} ${kid} ${kid} ${kid} ${kid} ${kid} ${kid} ${kid} ${kid} ${kid} ${kid} ${kid} ${kid} ${kid} ${kid} ${kid} ${kid} ${kid} ${kid} ${kid}`).then(f => {
@@ -149,14 +155,14 @@ client.on('message', msg => {
 });
 // mute/unmute
 client.on('message', msg => {
-    const muted = msg.guild.roles.cache.find(r => r.name === 'muted');
-    const mbr = msg.guild.roles.cache.find(r => r.name === 'Members');
+    const muted = msg.guild.roles.cache.find(r => r.name.toLowerCase() === 'muted');
+    const mbr = msg.guild.roles.cache.find(r => r.name.toLowerCase() === 'members');
     const kid = msg.mentions.members.first();
     const person = msg.guild.member(kid);
 
     if (msg.author.bot) return;
     if (msg.content.startsWith(`${prefix}mute`)) {
-        if (msg.member && msg.member.roles.cache.has('860401178009403393')) {
+        if (msg.member && msg.member.hasPermission('MUTE_MEMBERS')) {
             if (kid) {
                 person.roles.add(muted);
                 person.roles.remove(mbr);
@@ -174,7 +180,7 @@ client.on('message', msg => {
         }
     }
     if (msg.content.startsWith(`${prefix}unmute`)) {
-        if (msg.member && msg.member.roles.cache.has('860401178009403393')) {
+        if (msg.member && msg.member.hasPermission('MUTE_MEMBERS')) {
             if (kid) {
                 person.roles.remove(muted);
                 person.roles.add(mbr);
@@ -194,9 +200,11 @@ client.on('message', msg => {
 });
 // membercount 
 client.on('message', msg => {
+    let memberCount = msg.guild.members.cache.filter(member => !member.user.bot).size;
+    if (msg.author.bot) return;
     if (msg.content === `${prefix}membercount`) {
         const embed = new Discord.MessageEmbed()
-            .setTitle(`${msg.member.guild.memberCount} ***including bots***`)
+            .setTitle(`${memberCount}`)
             .setTimestamp()
             .setFooter('Made By Boomers Highness');
         msg.reply(embed);
@@ -204,15 +212,150 @@ client.on('message', msg => {
 });
 // about 
 client.on('message', msg => {
+    const kid = msg.mentions.members.first();
+    const args = msg.content.trim().split(/\s+/g);
+    const command = args.shift().toLowerCase();
+
     if (msg.author.bot) return;
-    if (msg.content === `${prefix}about`) {
-        const embed = new Discord.MessageEmbed()
-            .setTitle(`info about [${msg.author.tag}].`)
-            .setDescription(`\`account created at-[${msg.author.createdAt}]\``)
-            .setThumbnail(msg.author.displayAvatarURL({ dynamic: true }))
-            .setFooter(msg.author.id)
-            .setTimestamp();
-        msg.channel.send(embed);
+    if (command === `${prefix}about`) {
+        if (!kid) {
+            const embed = new Discord.MessageEmbed()
+                .setTitle(`info about [you].`)
+                .setAuthor(`nickname "${msg.member.nickname}"`)
+                .setImage(msg.author.displayAvatarURL({ size: 256, dynamic: true }))
+                .addFields(
+                    { name: 'account created on', value: `\`${msg.author.createdAt.toLocaleString()}\``},
+                    { name: 'account joined on', value: `\`${msg.member.joinedAt.toLocaleString()}\``},
+                    { name: 'boosted server since', value: `\`${msg.member.premiumSince}\``},
+                    { name: 'flags', value: `\`${msg.member.user.flags.toArray()}\`` },
+                    { name: 'roles', value: `${msg.member.roles.cache.array()}`, inline: true },
+                )
+                .setColor(`${msg.member.roles.highest.hexColor}`)
+                .setFooter(msg.author.id)
+                .setTimestamp();
+            msg.channel.send(embed).catch(err => console.error(err));
+        } else {
+            const embed = new Discord.MessageEmbed()
+                .setTitle(`info about [@${kid.user.tag}].`)
+                .setAuthor(`nickname "${kid.nickname}"`)
+                .setImage(kid.user.displayAvatarURL({ size: 256, dynamic: true }))
+                .addFields(
+                    { name: 'account created on', value: `\`${kid.user.createdAt.toLocaleString()}\`` },
+                    { name: 'account joined on', value: `\`${kid.joinedAt.toLocaleString()}\`` },
+                    { name: 'boosted server since', value: `\`${kid.premiumSince}\`` },
+                    { name: 'flags', value: `\`${kid.user.flags.toArray()}\`` },
+                    { name: 'roles', value: `${kid.roles.cache.array()}`, inline: true },
+                )
+                .setColor(`${kid.roles.highest.hexColor}`)
+                .setFooter(kid.user.id)
+                .setTimestamp();
+            msg.channel.send(embed).catch(err => console.error(err));
+        }
+    }
+});
+// kick
+client.on('message', msg => {
+    const kid = msg.mentions.users.first();
+    const member = msg.guild.member(kid);
+    if (msg.author.bot) return;
+    if (msg.content.startsWith(`${prefix}kick`)) {
+        if (msg.member && msg.member.hasPermission('KICK_MEMBERS')) {
+            if (member) {
+                member.kick().then(() =>{
+                    msg.reply(`kicked - ${member}`).then(E => {
+                        setTimeout(() => {
+                            E.delete();
+                            msg.delete();
+                        }, 5000);
+                    })
+                })
+            } else {
+                msg.reply('mention someone cuh').then(E => {
+                    setTimeout(() => {
+                        E.delete();
+                        msg.delete();
+                    }, 5000);
+                })
+            }
+        } else {
+            msg.reply('you sir stop right there dumbass, u cant use this command cuz u aint powerful enough ;)').then(E => {
+                setTimeout(() => {
+                    E.delete();
+                    msg.delete();
+                }, 5000);
+            })
+        }
+    }
+});
+//ban 
+// client.on('message', msg => {
+//     const kid = msg.mentions.users.first();
+//     const member = msg.guild.member(kid);
+//     if (msg.author.bot) return;
+//     if (msg.content.startsWith(`${prefix}ban`)) {
+//         if (msg.member && msg.member.hasPermission('BAN_MEMBERS')) {
+//             if (member) {
+//                 member.ban().then(() => {
+//                     msg.reply(`banned - ${member}`).then(E => {
+//                         setTimeout(() => {
+//                             E.delete();
+//                             msg.delete();
+//                         }, 5000);
+//                     })
+//                 })
+//             } else {
+//                 msg.reply('mention someone cuh').then(E => {
+//                     setTimeout(() => {
+//                         E.delete();
+//                         msg.delete();
+//                     }, 5000);
+//                 })
+//             }
+//         } else {
+//             msg.reply('you sir stop right there dumbass, u cant use this command cuz u aint powerful enough ;)').then(E => {
+//                 setTimeout(() => {
+//                     E.delete();
+//                     msg.delete();
+//                 }, 5000);
+//             })
+//         }
+//     }
+//     if (msg.content.startsWith(`${prefix}unban`)) {
+//         if (msg.member && msg.member.hasPermission('BAN_MEMBERS')) {
+//             if (member) {
+//                msg.guild.fetchBans().then(penis =>{
+                   
+//                })
+//             } else {
+//                 msg.reply('mention someone cuh').then(E => {
+//                     setTimeout(() => {
+//                         E.delete();
+//                         msg.delete();
+//                     }, 3000);
+//                 })
+//             }
+//         } else {
+//                 msg.reply('you sir stop right there dumbass, u cant use this command cuz u aint powerful enough ;)').then(E => {
+//                     setTimeout(() => {
+//                         E.delete();
+//                         msg.delete();
+//                 }, 5000);
+//             })
+//         }
+//     }
+// });
+
+// id show
+client.on('message', msg => {
+    const kid = msg.mentions.users.first();
+    const member = msg.guild.member(kid);
+    if (msg.author.bot) return;
+    if (msg.content.startsWith(`${prefix}id`)) {
+        if (member) {
+            msg.reply(`this is ${member}'s id - \`${member.user.id}\``);
+        } else {
+            msg.reply(`\`${msg.member.valueOf()}\``);
+        }
     }
 });
 client.on('messageDelete', message => {
@@ -229,15 +372,15 @@ client.on('messageUpdate', message => {
 // logs
 client.on('messageDelete', msgdel =>{
     if (msgdel.author.bot) return;
-    console.log(`${msgdel.author.username}[${msgdel.author}] deleted - ["${msgdel.content}"] - [${msgdel.guild.name}[id-${msgdel.guild.id}] | channel [${msgdel.channel.name}] | ${msgdel.createdAt.toLocaleTimeString()} | ${msgdel.createdAt.toLocaleDateString()}] [deleted message]`);
+    console.log(`${msgdel.author.tag}[${msgdel.author}] deleted - ["${msgdel.content}"] - [${msgdel.guild.name}[id-${msgdel.guild.id}] | channel [${msgdel.channel.name}] | ${msgdel.createdAt.toLocaleTimeString()} | ${msgdel.createdAt.toLocaleDateString()}] [deleted message]`);
 });
 client.on('messageUpdate', (newMessage, oldMessage) =>{
     if (oldMessage.author.bot) return;
-    console.log(`${newMessage.author.username}[${newMessage.author}] edited - before[${newMessage.content}] after[${oldMessage.content}] - [${newMessage.guild.name}[id-${newMessage.guild.id}] | channel [${newMessage.channel.name}] | ${newMessage.createdAt.toLocaleTimeString()} | ${newMessage.createdAt.toLocaleDateString()}] [edited message]`)
+    console.log(`${newMessage.author.tag}[${newMessage.author}] edited - before[${newMessage.content}] after[${oldMessage.content}] - [${newMessage.guild.name}[id-${newMessage.guild.id}] | channel [${newMessage.channel.name}] | ${newMessage.createdAt.toLocaleTimeString()} | ${newMessage.createdAt.toLocaleDateString()}] [edited message]`)
 });
 client.on('message', msg => {
     if (msg.author.bot) return;
-    console.log(`${msg.author.username}[${msg.author}] said ["${msg.content}"] in ${msg.guild.name}[id-${msg.guild.id}] | channel [${msg.channel.name}] | ${msg.createdAt.toLocaleTimeString()} | ${msg.createdAt.toLocaleDateString()}] [message]`)
+    console.log(`${msg.author.tag}[${msg.author}] said ["${msg.content}"] in ${msg.guild.name}[id-${msg.guild.id}] | channel [${msg.channel.name}] | ${msg.createdAt.toLocaleTimeString()} | ${msg.createdAt.toLocaleDateString()}] [message]`)
 });
 // end of logs ^
 
@@ -246,18 +389,20 @@ client.on('message', msg => {
 client.on('message', msg => {
     if (msg.author.bot) return;
     if (msg.content.includes('<@!420871370516201473>')){
-            msg.react('🖕')
+        if (msg.author.id === '420871370516201473') {
+            msg.react('👍');
+        } else {
+            msg.react('🖕');
         }
+    }
 });
 // no links n3rd
 client.on('message', msg =>{
-    const server = client.guilds.cache.get('860397277306814535');
-    if (!server) return;
-    if (msg.author.id === '420871370516201473') return;
-    if (msg.member  &&  msg.member.roles.cache.has("865355545044844604")) {
+    if (msg.author.id === '420871370516201473' || msg.author.id === '788646324219281410') return;
+    if (msg.member && msg.member.hasPermission('ADMINISTRATOR') || msg.member && msg.member.roles.cache.has('865355545044844604')) {
        return;
     } else {
-        if (msg.content.includes("https://")) {
+        if (msg.content.includes("https://") || msg.content.includes("discord.gg/") || msg.content.includes("http://")) {
             msg.reply('no links dummie').then(E => {
                 setTimeout(() => {
                     E.delete()
@@ -265,14 +410,6 @@ client.on('message', msg =>{
             })
             msg.delete().catch(err => console.error(err));
         }
-        if (msg.content.includes("http://")) {
-            msg.reply('no links dummie').then(E => {
-                setTimeout(() => {
-                    E.delete()
-                }, 2000);
-            })
-            msg.delete().catch(err => console.error(err));
-        }   
     }
 });
 // end of no links
@@ -281,8 +418,3 @@ client.on('message', msg =>{
     if (msg.content === 'f') return msg.react('🇫');
 });
 // end of memes
-
-
-
-
-
